- special rollback for transactional server (bsc#1172273)

+++ snapper/Enum.h
@@ -63,7 +63,7 @@ namespace snapper
 	if ((size_t)(value) >= names.size())
 	    throw OutOfRangeException();
 
-	return names[value];
+	return names[(size_t)(value)];
     }
